fix(backend): move r2 trash helpers under _backend for edge worktrees
Edge functions cannot import repo-root scripts/ in CI supabase worktrees.
Move shared trash utilities to utils/r2_trash_shared.ts and re-export from
scripts/r2_trash_utils.ts for ops tooling.
