+++ supabase/functions/_backend/utils/s3.ts
@@ -1,8 +1,8 @@
 import type { Context } from 'hono'
 import type { Database } from '../utils/supabase.types.ts'
 import { S3Client } from '@bradenmacdonald/s3-lite-client'
-import type { RawS3LiteClient } from '../../../../scripts/r2_trash_utils.ts'
-import { conditionalDeleteSource, encodeS3LiteCopySourceKey, isObjectNotFoundError } from '../../../../scripts/r2_trash_utils.ts'
+import type { RawS3LiteClient } from './r2_trash_shared.ts'
+import { conditionalDeleteSource, encodeS3LiteCopySourceKey, isObjectNotFoundError } from './r2_trash_shared.ts'
 import { cloudlog, cloudlogErr, serializeError } from './logging.ts'
 import { getManifestStorageCandidateKeys } from './manifest_encoding.ts'
 import { getEnv } from './utils.ts'
